Refactor statistic endpoints into separate files under one module
This will make testing easier, as we can add the endpoint tests
into each file instead of making a mammoth statistic testing file.
