+++ src/stats/forward_destinations.rs
@@ -1,0 +1,59 @@
+/* Pi-hole: A black hole for Internet advertisements
+*  (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+*  Network-wide ad blocking via your own hardware.
+*
+*  API
+*  Forward Destinations API Endpoint
+*
+*  This file is copyright under the latest version of the EUPL.
+*  Please see LICENSE file for your rights under this license. */
+
+use ftl::FtlConnectionType;
+use rmp::decode::DecodeStringError;
+use rmp::Marker;
+use rocket::State;
+use std::collections::HashMap;
+use util;
+
+/// Get the forward destinations
+#[get("/stats/forward_destinations")]
+pub fn forward_destinations(ftl: State<FtlConnectionType>) -> util::Reply {
+    let mut con = ftl.connect("forward-dest")?;
+
+    // Create a 4KiB string buffer
+    let mut str_buffer = [0u8; 4096];
+    let mut forward_destinations: HashMap<String, f32> = HashMap::new();
+
+    loop {
+        // Read in the hostname, unless we are at the end of the list
+        let name = match con.read_str(&mut str_buffer) {
+            Ok(name) => name.to_string(),
+            Err(e) => {
+                // Check if we received the EOM
+                if let DecodeStringError::TypeMismatch(marker) = e {
+                    if marker == Marker::Reserved {
+                        // Received EOM
+                        break;
+                    }
+                }
+
+                // Unknown read error
+                return util::reply_error(util::Error::Unknown);
+            }
+        };
+
+        let ip = con.read_str(&mut str_buffer)?;
+        let percentage = con.read_f32()?;
+
+        // The key will be `hostname|IP` if the hostname exists, otherwise just the IP address
+        let key = if ip.len() > 0 {
+            format!("{}|{}", name, ip)
+        } else {
+            name
+        };
+
+        forward_destinations.insert(key, percentage);
+    }
+
+    util::reply_data(forward_destinations)
+}
