+++ src/stats/mod.rs
@@ -1,0 +1,37 @@
+/* Pi-hole: A black hole for Internet advertisements
+*  (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+*  Network-wide ad blocking via your own hardware.
+*
+*  API
+*  Statistic API Endpoints
+*
+*  This file is copyright under the latest version of the EUPL.
+*  Please see LICENSE file for your rights under this license. */
+
+mod summary;
+mod top_domains;
+mod top_clients;
+mod forward_destinations;
+mod query_types;
+mod history;
+mod recent_blocked;
+mod clients;
+mod unknown_queries;
+mod over_time_history;
+mod over_time_forward_destinations;
+mod over_time_query_types;
+mod over_time_clients;
+
+pub use self::summary::*;
+pub use self::top_domains::*;
+pub use self::top_clients::*;
+pub use self::forward_destinations::*;
+pub use self::query_types::*;
+pub use self::history::*;
+pub use self::recent_blocked::*;
+pub use self::clients::*;
+pub use self::unknown_queries::*;
+pub use self::over_time_history::*;
+pub use self::over_time_forward_destinations::*;
+pub use self::over_time_query_types::*;
+pub use self::over_time_clients::*;
