+++ src/stats/over_time_clients.rs
@@ -1,0 +1,62 @@
+/* Pi-hole: A black hole for Internet advertisements
+*  (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+*  Network-wide ad blocking via your own hardware.
+*
+*  API
+*  Clients Over Time API Endpoint
+*
+*  This file is copyright under the latest version of the EUPL.
+*  Please see LICENSE file for your rights under this license. */
+
+use ftl::FtlConnectionType;
+use rmp::decode::ValueReadError;
+use rmp::Marker;
+use rocket::State;
+use std::collections::HashMap;
+use util;
+
+/// Get the client queries over time
+#[get("/stats/overTime/clients")]
+pub fn over_time_clients(ftl: State<FtlConnectionType>) -> util::Reply {
+    let mut con = ftl.connect("ClientsoverTime")?;
+
+    let mut over_time: HashMap<i32, Vec<i32>> = HashMap::new();
+
+    loop {
+        // Get the timestamp, unless we are at the end of the list
+        let timestamp = match con.read_i32() {
+            Ok(timestamp) => timestamp,
+            Err(e) => {
+                // Check if we received the EOM
+                if let ValueReadError::TypeMismatch(marker) = e {
+                    if marker == Marker::Reserved {
+                        // Received EOM
+                        break;
+                    }
+                }
+
+                // Unknown read error
+                return util::reply_error(util::Error::Unknown);
+            }
+        };
+
+        // Create a new step in the graph (stores the value of each client usage at that time)
+        let mut step = Vec::new();
+
+        // Get all the data for this step
+        loop {
+            let client = con.read_i32()?;
+
+            // Marker for the end of this step
+            if client == -1 {
+                break;
+            }
+
+            step.push(client);
+        }
+
+        over_time.insert(timestamp, step);
+    }
+
+    util::reply_data(over_time)
+}
